added fish updates

+++ compiled/functions/fish/classes/Fish.js
@@ -1,8 +1,8 @@
 import { ActionRowBuilder, ButtonBuilder, ButtonStyle, Collection, ComponentType, EmbedBuilder, } from "discord.js";
 import { createJIDeck, formatCardSideways, formatDeck, } from "../../cards/basic_card_functions.js";
 import { colors, deleteSoon, hasAdminForGames, msToRelTimestamp, removeC, replyThenDelete, } from "../../util.js";
-import FishPlayerCollection from "./FishPlayerCollection.js";
-import { CardsPerHalfSuit, customIds, customIdValues, gameLength, timeToMakeCall, } from "../fish_types.js";
+import FishPlayerCollection, { TeamsAreDisjointError, } from "./FishPlayerCollection.js";
+import { CardsPerHalfSuit, customIds, customIdValues, gameLength, timeToChooseHalfSuitToMakeCall, } from "../fish_types.js";
 import { deckHasCard, extrapolateCard, formatHalfSuitCall, halfSuitSelectMenuRow, hasOtherCardInSameHalfSuit, nonDisjointCards, } from "../fish_functions.js";
 import { channelsWithActiveGames } from "../../../main.js";
 const callButton = new ButtonBuilder()
@@ -12,7 +12,7 @@ const callButton = new ButtonBuilder()
 const viewCardsButton = new ButtonBuilder()
     .setCustomId(customIds.viewCards)
     .setLabel("View Cards")
-    .setStyle(ButtonStyle.Secondary);
+    .setStyle(ButtonStyle.Primary);
 const viewTableButton = new ButtonBuilder()
     .setCustomId(customIds.viewTable)
     .setLabel("Table")
@@ -29,6 +29,7 @@ const abortDJButton = new ButtonBuilder()
     .setCustomId(customIds.abortDJ)
     .setLabel("Abort")
     .setStyle(ButtonStyle.Danger);
+const bottomRowWoDisjoint = new ActionRowBuilder().addComponents(viewCardsButton, viewTableButton, callButton);
 const bottomRow = new ActionRowBuilder().addComponents(viewCardsButton, viewTableButton, callButton, disjointButton);
 function idToAskId(id) {
     return `ask_fish${id}`;
@@ -58,9 +59,10 @@ function callCidToId(customId) {
     return customId.slice(9);
 }
 export default class Fish {
-    constructor(channel, hostId, team0, team1) {
+    constructor(channel, hostId, team0, team1, deleteHistory) {
         this.channel = channel;
         this.hostId = hostId;
+        this.deleteHistory = deleteHistory;
         this.ongoingAskId = null;
         this.ongoingAskMessage = null;
         this.ongoingCall = false;
@@ -94,6 +96,20 @@ export default class Fish {
                 new EmbedBuilder()
                     .setAuthor({ name: "Fish" })
                     .setTitle(`Team ${winner + 1} has won!`)
+                    .addFields([
+                    {
+                        name: "Players",
+                        value: this.players.formatTeamNamesOnly(),
+                    },
+                    {
+                        name: "Team 1 (Final Table)",
+                        value: this.players.formatTeamTableCards(0),
+                    },
+                    {
+                        name: "Team 2 (Final Table)",
+                        value: this.players.formatTeamTableCards(1),
+                    },
+                ])
                     .setColor(colors.green),
             ],
             components: [],
@@ -104,7 +120,7 @@ export default class Fish {
         const opponents = this.players.opponents(player.id);
         for (const opponent of opponents.values()) {
             const button = new ButtonBuilder()
-                .setLabel(opponent.username)
+                .setLabel(opponent.user.username)
                 .setCustomId(useDisjointCodes
                 ? idToDisjointId(opponent.id)
                 : idToAskId(opponent.id))
@@ -120,7 +136,7 @@ export default class Fish {
         const team = teamNumber === 0 ? this.players.team0 : this.players.team1;
         for (const player of team.values()) {
             row.addComponents(new ButtonBuilder()
-                .setLabel(player.username)
+                .setLabel(player.user.username)
                 .setCustomId(idToCallCid(player.id))
                 .setStyle(ButtonStyle.Secondary));
         }
@@ -145,7 +161,7 @@ export default class Fish {
             try {
                 const buttonInteraction = await msg.awaitMessageComponent({
                     componentType: ComponentType.Button,
-                    time: timeToMakeCall,
+                    time: timeToChooseHalfSuitToMakeCall,
                     filter: i => i.user.id === player.id && cIdIsCallFish(i.customId),
                     idle: 0,
                 });
@@ -228,7 +244,7 @@ export default class Fish {
         this.players.removeSuitFromPlayers(suit);
     }
     async callCollect(buttonInteraction) {
-        const timeUp = msToRelTimestamp(timeToMakeCall);
+        const timeUp = msToRelTimestamp(timeToChooseHalfSuitToMakeCall);
         const msg = await buttonInteraction.reply({
             embeds: [
                 new EmbedBuilder()
@@ -239,7 +255,7 @@ export default class Fish {
                     .setTitle("Call")
                     .setDescription(`Please use the menu below to select a half suit to call ${timeUp}.
 Once you select a half suit, you will be prompted to select the player in your team who has each card in that half suit.
-If you wish to abort the call, press the "Abort" button.`)
+If you wish to cancel the call, let it time out.`)
                     .setColor(colors.blurple),
             ],
             components: [halfSuitSelectMenuRow],
@@ -249,7 +265,7 @@ If you wish to abort the call, press the "Abort" button.`)
         msg
             .awaitMessageComponent({
             componentType: ComponentType.StringSelect,
-            time: timeToMakeCall,
+            time: timeToChooseHalfSuitToMakeCall,
             filter: i => i.customId === customIds.callFishSelect &&
                 i.user.id === buttonInteraction.user.id,
         })
@@ -278,7 +294,7 @@ ${formatDeck(CardsPerHalfSuit[selectInteraction.values[0]])}`,
             return;
         }
         this.ongoingAskId = playerId;
-        const timeUp = msToRelTimestamp(timeToMakeCall);
+        const timeUp = msToRelTimestamp(timeToChooseHalfSuitToMakeCall);
         this.ongoingAskMessage = await buttonInteraction.reply({
             content: `Please type the card you wish to ask <@${playerId}> for ${timeUp}.`,
             fetchReply: true,
@@ -302,7 +318,7 @@ ${formatDeck(CardsPerHalfSuit[selectInteraction.values[0]])}`,
         });
         try {
             const newBi = await msg.awaitMessageComponent({
-                time: timeToMakeCall,
+                time: timeToChooseHalfSuitToMakeCall,
                 idle: 0,
                 componentType: ComponentType.Button,
                 filter: i => {
@@ -334,13 +350,26 @@ ${formatDeck(CardsPerHalfSuit[selectInteraction.values[0]])}`,
                 });
             }
             else {
-                await this.channel.send({
-                    content: `:red_circle: ${player}, you are not disjoint with ${callWith}. You will give them the following cards in order to be disjoint. 
-${formatDeck(matchingCards)}
+                const formattedMatchingDeck = formatDeck(matchingCards);
+                await Promise.all([
+                    this.channel.send({
+                        content: `:red_circle: ${player}, you are not disjoint with ${callWith}. You will give them **${matchingCards.length} cards** in order to be disjoint. 
 Now, you may no longer ask each other for cards.`,
-                });
+                    }),
+                    player.user.send({
+                        content: `You were not disjoint with ${callWith}. You gave them the following cards:
+${formattedMatchingDeck}`,
+                    }),
+                    callWith.user.send({
+                        content: `You are now disjoint with ${player}. You received the following cards:
+${formattedMatchingDeck}
+Here is the ${player}'s original hand:
+${player.formatHand()}`,
+                    }),
+                ]);
                 player.hand = nonMatchingCards;
                 callWith.hand.push(...matchingCards);
+                callWith.sortHand();
             }
             player.disjoint.push(callWith.id);
             callWith.disjoint.push(player.id);
@@ -350,11 +379,25 @@ Now, you may no longer ask each other for cards.`,
             msg.delete();
         }
         if (player.out) {
-            this.players.setFirstAvailablePlayer();
-            await this.channel.send({
-                content: `${player}, you are now disjoint with all members of the opposing team. The turn will pass to <@${this.players.currentPlayerId}>.`,
-            });
-            await this.turn();
+            try {
+                this.players.setFirstAvailablePlayer(player.team);
+                await this.channel.send({
+                    content: `${player}, you are now disjoint with all members of the opposing team. The turn will pass to <@${this.players.currentPlayerId}>.`,
+                });
+                await this.turn();
+            }
+            catch (e) {
+                if (e instanceof TeamsAreDisjointError) {
+                    await this.allDisjointTurn();
+                }
+                else {
+                    this.endCollectors();
+                    this.channel.send({
+                        content: "An unexpected error occurred and the game has been aborted.",
+                    });
+                    console.error(e);
+                }
+            }
         }
     }
     async buttonCollect(buttonInteraction) {
@@ -363,16 +406,20 @@ Now, you may no longer ask each other for cards.`,
                 embeds: [
                     new EmbedBuilder()
                         .setTitle("Game Info")
-                        .addFields({
-                        name: "Players",
-                        value: this.players.formatTeamsAndHandLengths(),
-                    }, {
-                        name: "Team 1",
-                        value: this.players.formatTeamTableCards(0),
-                    }, {
-                        name: "Team 2",
-                        value: this.players.formatTeamTableCards(1),
-                    })
+                        .addFields([
+                        {
+                            name: "Players",
+                            value: this.players.formatTeamsAndHandLengths(),
+                        },
+                        {
+                            name: "Team 1",
+                            value: this.players.formatTeamTableCards(0),
+                        },
+                        {
+                            name: "Team 2",
+                            value: this.players.formatTeamTableCards(1),
+                        },
+                    ])
                         .setColor(colors.green),
                 ],
                 ephemeral: true,
@@ -393,6 +440,7 @@ Now, you may no longer ask each other for cards.`,
                     content: "You cannot call while someone is trying to ask someone for a card.",
                     ephemeral: true,
                 });
+                return;
             }
             await this.callCollect(buttonInteraction);
             return;
@@ -433,9 +481,10 @@ Now, you may no longer ask each other for cards.`,
         const turnMessageRef = this.currentTurnMessage;
         if (this.currentTurnMessage) {
             this.currentTurnMessage.edit({ components: [] });
-            setTimeout(() => {
-                turnMessageRef.delete();
-            }, 60000);
+            if (this.deleteHistory)
+                setTimeout(() => {
+                    turnMessageRef.delete();
+                }, 60000);
         }
         this.ongoingAskId = null;
         this.ongoingAskMessage = null;
@@ -443,6 +492,14 @@ Now, you may no longer ask each other for cards.`,
         this.currentTurnMessage = await this.channel.send({
             content: `<@${this.players.currentPlayerId}>, it is your turn.`,
             components: [row, bottomRow],
+        });
+    }
+    async allDisjointTurn() {
+        this.ongoingAskId = null;
+        this.ongoingAskMessage = null;
+        this.currentTurnMessage = await this.channel.send({
+            content: `Both teams are disjoint with each other. You may only make calls now.`,
+            components: [bottomRowWoDisjoint],
         });
     }
     async askCollect(msg) {
@@ -467,7 +524,7 @@ Please try asking again.`);
         if (!targetPlayer)
             return;
         if (deckHasCard(targetPlayer.hand, card)) {
-            this.players.currentPlayer.hand.push(card);
+            this.players.currentPlayer.insertCard(card);
             removeC(targetPlayer.hand, x => x.suit === card.suit && x.value === card.value);
             const content = `:green_circle: ${targetPlayer} had the ${formatCardSideways(card)}. They give it to ${this.players.currentPlayer}.`;
             this.ongoingAskMessage
@@ -498,9 +555,15 @@ Please try asking again.`);
             hasAdminForGames(msg.author.id, msg.member.permissions, this.hostId)) {
             this.endCollectors();
             msg.reply("Game aborted.");
+            return;
         }
         if (this.ongoingAskId) {
             this.askCollect(msg);
+        }
+        if (this.deleteHistory && msg.author.id !== msg.client.user.id) {
+            setTimeout(() => {
+                msg.delete();
+            }, 120000);
         }
     }
     async gameLogic() {
